+++ pydomcfg/domzgr/zgr.py
@@ -2,6 +2,7 @@
 Base class to generate NEMO v4.0 vertical grids.
 """
 
+
 from typing import Tuple, Union
 
 import xarray as xr