+++ pydomcfg/tests/test_zco.py
@@ -2,6 +2,8 @@
 Tests for zco
 """
 
+from io import StringIO
+
 import numpy as np
 import pytest
 import xarray as xr
@@ -9,10 +11,11 @@ import xarray as xr
 import pydomcfg  # noqa: F401
 
 from .bathymetry import Bathymetry
-from .data import ORCA2_VGRID
+from .data import NML_REF_PATH, ORCA2_NAMELIST, ORCA2_VGRID
 
 
-def test_zco_orca2():
+@pytest.mark.parametrize("from_namelist", (True, False))
+def test_zco_orca2(from_namelist):
     """
     The test consists in reproducing ORCA2 grid
     z3T/W and e3T/W as computed by NEMO v3.6
@@ -24,22 +27,31 @@ def test_zco_orca2():
     # Bathymetry dataset
     ds_bathy = Bathymetry(1.0e3, 1.2e3, 1, 1).flat(5.0e3)
 
-    # Set number of vertical levels
-    ds_bathy.domcfg.jpk = 31
+    if from_namelist:
+        pytest.importorskip("f90nml")
 
-    # zco mesh with analytical e3 using ORCA2 input parameters
-    # See pag 62 of v3.6 manual for the input parameters
-    dsz_an = ds_bathy.domcfg.zco(
-        ppdzmin=10.0,
-        pphmax=5000.0,
-        ppkth=21.43336197938,
-        ppacr=3,
-        ppsur=-4762.96143546300,
-        ppa0=255.58049070440,
-        ppa1=245.58132232490,
-        ldbletanh=False,
-        ln_e3_dep=False,
-    )
+        # Set reference namelist
+        ds_bathy.domcfg.nml_ref_path = NML_REF_PATH
+
+        # Infer parameters from namelist
+        dsz_an = ds_bathy.domcfg.from_namelist(StringIO(ORCA2_NAMELIST))
+
+    else:
+        # Set number of vertical levels
+        ds_bathy.domcfg.jpk = 31
+
+        # zco mesh with analytical e3 using ORCA2 input parameters
+        # See pag 62 of v3.6 manual for the input parameters
+        dsz_an = ds_bathy.domcfg.zco(
+            ppdzmin=10.0,
+            pphmax=5000.0,
+            ppkth=21.43336197938,
+            ppacr=3,
+            ppsur=-4762.96143546300,
+            ppa0=255.58049070440,
+            ppa1=245.58132232490,
+            ln_e3_dep=False,
+        )
 
     # reference ocean.output values are
     # given with 4 digits precision
@@ -103,15 +115,11 @@ def test_zco_errors():
     ds_bathy = Bathymetry(1.0e3, 1.2e3, 1, 1).flat(5.0e3)
     ds_bathy.domcfg.jpk = 10
 
-    # Without ldbletanh flag, only allow all pps set or none of them
+    # Only allow all pps set or none of them
     with pytest.raises(
         ValueError, match="ppa2, ppkth2 and ppacr2 MUST be all None or all float"
     ):
         ds_bathy.domcfg.zco(**kwargs, ppa2=1, ppkth2=1, ppacr2=None)
-
-    # When ldbletanh flag is True, all coefficients must be specified
-    with pytest.raises(ValueError, match="ppa2, ppkth2 and ppacr2 MUST be all float"):
-        ds_bathy.domcfg.zco(**kwargs, ldbletanh=True, ppa2=1, ppkth2=1, ppacr2=None)
 
 
 def test_zco_warnings():
@@ -123,20 +131,11 @@ def test_zco_warnings():
 
     # Uniform: Ignore stretching
     kwargs = dict(ppdzmin=10, pphmax=5.0e3, ppkth=0, ppacr=0)
-    expected = ds_bathy.domcfg.zco(**kwargs, ppsur=None, ppa0=999_999, ppa1=None)
+    expected = ds_bathy.domcfg.zco(**kwargs, ppsur=None, ppa0=None, ppa1=None)
     with pytest.warns(
         UserWarning, match="ppsur, ppa0 and ppa1 are ignored when ppacr == ppkth == 0"
     ):
         actual = ds_bathy.domcfg.zco(**kwargs, ppsur=2, ppa0=2, ppa1=2)
-    xr.testing.assert_identical(expected, actual)
-
-    # ldbletanh OFF: Ignore double tanh
-    kwargs = dict(ppdzmin=10, pphmax=5.0e3, ldbletanh=False)
-    expected = ds_bathy.domcfg.zco(**kwargs, ppa2=None, ppkth2=None, ppacr2=None)
-    with pytest.warns(
-        UserWarning, match="ppa2, ppkth2 and ppacr2 are ignored when ldbletanh is False"
-    ):
-        actual = ds_bathy.domcfg.zco(**kwargs, ppa2=2, ppkth2=2, ppacr2=2)
     xr.testing.assert_identical(expected, actual)
 
     # Uniform case: Ignore double tanh